+++ .env.copy
@@ -12,3 +12,4 @@ MAX_SELL_RETRIES=5
 AUTO_SELL_DELAY=1000
 LOG_LEVEL=info
 MIN_POOL_SIZE=10
+PAPER_TRADE=true
